refactor: 멤버 운동 엔티티를 exercise 도메인으로 이동
