+++ src/main/java/umc/cockple/demo/domain/exercise/converter/command/ExerciseParticipationCommandMapper.java
@@ -6,7 +6,7 @@ import umc.cockple.demo.domain.exercise.domain.Guest;
 import umc.cockple.demo.domain.exercise.dto.participation.ExerciseCancelDTO;
 import umc.cockple.demo.domain.exercise.dto.participation.ExerciseJoinDTO;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.domain.MemberExercise;
+import umc.cockple.demo.domain.exercise.domain.MemberExercise;
 
 @Component
 public class ExerciseParticipationCommandMapper {
