+++ src/main/java/umc/cockple/demo/domain/exercise/repository/ExerciseRepositoryCustomImpl.java
@@ -10,7 +10,7 @@ import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.domain.QExercise;
 import umc.cockple.demo.domain.exercise.domain.QExerciseAddr;
 import umc.cockple.demo.domain.exercise.repository.support.ExerciseRecommendationSearchCondition;
-import umc.cockple.demo.domain.member.domain.QMemberExercise;
+import umc.cockple.demo.domain.exercise.domain.QMemberExercise;
 import umc.cockple.demo.domain.member.domain.QMemberParty;
 import umc.cockple.demo.domain.party.domain.QParty;
 import umc.cockple.demo.domain.party.domain.QPartyImg;
