+++ src/main/java/umc/cockple/demo/domain/exercise/repository/MemberExerciseRepository.java
@@ -1,12 +1,12 @@
-package umc.cockple.demo.domain.member.repository;
+package umc.cockple.demo.domain.exercise.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
+import umc.cockple.demo.domain.exercise.domain.MemberExercise;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.domain.MemberExercise;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
