+++ src/main/java/umc/cockple/demo/domain/exercise/service/support/reader/ExerciseParticipantReader.java
@@ -7,8 +7,8 @@ import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.exception.ExerciseErrorCode;
 import umc.cockple.demo.domain.exercise.exception.ExerciseException;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.domain.MemberExercise;
-import umc.cockple.demo.domain.member.repository.MemberExerciseRepository;
+import umc.cockple.demo.domain.exercise.domain.MemberExercise;
+import umc.cockple.demo.domain.exercise.repository.MemberExerciseRepository;
 
 import java.util.Collections;
 import java.util.HashSet;
