+++ src/test/java/umc/cockple/demo/support/fixture/MemberFixture.java
@@ -1,7 +1,7 @@
 package umc.cockple.demo.support.fixture;
 
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.domain.MemberExercise;
+import umc.cockple.demo.domain.exercise.domain.MemberExercise;
 import umc.cockple.demo.domain.member.domain.MemberParty;
 import umc.cockple.demo.domain.member.enums.MemberPartyStatus;
 import umc.cockple.demo.domain.member.enums.MemberStatus;
